refactor sensors logic
