+++ include/http/server.h
@@ -6,7 +6,7 @@
 #include <ESP8266HTTPClient.h>
 #include <ESP8266WebServer.h>
 #include <ESP8266mDNS.h>
-#include "pzems/pzem.h"
+#include "sensors/sensor.h"
 #include "utils/pin.h"
 #include "utils/wifi.h"
