+++ src/sensors/classes/dcpzem.cpp
@@ -1,4 +1,4 @@
-#include "pzems/classes/dcpzem.h"
+#include "sensors/classes/dcpzem.h"
 
 // Public
